fix _remove_zipfolder

+++ backend_web/app/Services/Open/PdftojpgService.php
@@ -41,7 +41,7 @@ class PdftojpgService extends BaseService
     private function _remove_zipfolder()
     {
         $pathfolder = $this->pathdown."/$this->foldername";
-        rmdir($pathfolder);
+        if(is_dir($pathfolder)) rmdir($pathfolder);
     }
 
     private function _exec($cmd)
